fix type in calculateVatForTransaction (return NUMERIC)

+++ mini_accountant/routines/calculateVatForTransaction.sql
@@ -1,10 +1,4 @@
-CREATE OR REPLACE FUNCTION `invoicemaker-f5e1d.mini_accountant.calculateVatForTransaction`(
-    transactionId INT64,
-    vat NUMERIC,
-    transactions ARRAY < STRUCT < DATETIME TIMESTAMP,
-    SUM NUMERIC,
-    EXCHANGE_RATE NUMERIC >>
-  ) RETURNS FLOAT64 LANGUAGE js AS R """
+CREATE OR REPLACE FUNCTION `invoicemaker-f5e1d.mini_accountant.calculateVatForTransaction`(transactionId INT64, vat NUMERIC, transactions ARRAY<STRUCT<DATETIME TIMESTAMP, SUM NUMERIC, EXCHANGE_RATE NUMERIC>>) RETURNS NUMERIC LANGUAGE js AS R"""
 function calculateVatForTransaction(transactionId, vat, transactions) {
   let vatRemains = vat;
   for(let i = 0; i < transactions.length; i++){
